Updated README file commit
simple_shell

A simple command-line prompt which takes the basics of commands that are present in the bash shell and runs them. This shell was built as a project for ALX software engineering School.

Description

simple_shell is a simple UNIX command language interpreter that reads commands from either a file or standard input and executes them.

Invocation

To invoke Simple_shell, compile all .c files in the repository, and run the resulting executable file using:
gcc *.c -o filename

Simple_shell can be invoked both interactively and non-interactively. If Simple_shell is invoked with a standard input not connected to a terminal, it reads and executes received commands in order.

Example:

$ echo "echo 'hello'" | ./filename 'hello' $

Authors
The authors of this simple shell project are Hftamu-S and Ayele-Getu-Gebremariyam